Pass 'UploadNewAlbum' component as children to the 'AlbumsGrid'
